installed react-router and setup routing
