+++ src/Components/Navbar/Navbar.jsx
@@ -4,6 +4,7 @@ import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
+import {Link} from 'react-router-dom'
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -28,10 +29,10 @@ export default function Navbar() {
           <Typography variant="h6" className={classes.title}>
             Bookmarked
           </Typography>
-          <Button color="inherit">Home</Button>
-          <Button color="inherit">About</Button>
-          <Button color="inherit">SignUp</Button>
-          <Button color="inherit">Login</Button>
+        <Link to="/">  <Button color="inherit">Home</Button></Link>
+        <Link to="/about"> <Button color="inherit">About</Button></Link>
+        <Link to="/signup"> <Button color="inherit">SignUp</Button></Link>
+        <Link to="/login"><Button color="inherit">Login</Button></Link>
         </Toolbar>
       </AppBar>
     </div>
