+++ src/Pages/About/About.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import Navbar from '../../Components/Navbar/Navbar';
+
+function About() {
+    return (
+        <>
+        <Navbar/>
+        <h1>About</h1>
+        </>
+    )
+}
+
+export default About
