+++ src/Pages/Dashboard/Dashboard.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Dashboard() {
+    return (
+        <div>
+          <h1>Dashboard</h1>  
+        </div>
+    )
+}
+
+export default Dashboard
