+++ src/Pages/Error/Error.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Error() {
+    return (
+        <div>
+            <h1>Sorry , The page could not be found</h1>
+        </div>
+    )
+}
+
+export default Error
