+++ src/Pages/Profile/Profile.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Profile() {
+    return (
+        <div>
+            <h1>Profile</h1>
+        </div>
+    )
+}
+
+export default Profile
